requirements.txt
requests==2.31.0
gradio==4.44.0
python-dotenv==1.0.0
tqdm==4.66.1
urllib3==2.0.7